+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/viewmodel/messages/MessageComposerViewModelTest.kt
@@ -445,6 +445,7 @@ internal class MessageComposerViewModelTest {
                         maxAttachmentCount = maxAttachmentCount,
                     ),
                     channelState = MutableStateFlow(channelState),
+                    globalState = MutableStateFlow(globalState),
                 ),
             )
         }