+++ stream-chat-android-ui-common/src/test/kotlin/io/getstream/chat/android/ui/common/feature/messages/composer/MessageComposerControllerTests.kt
@@ -25,10 +25,8 @@ import io.getstream.chat.android.models.ChannelData
 import io.getstream.chat.android.models.Command
 import io.getstream.chat.android.models.Config
 import io.getstream.chat.android.models.DraftMessage
-import io.getstream.chat.android.models.InitializationState
 import io.getstream.chat.android.models.Member
 import io.getstream.chat.android.models.User
-import io.getstream.chat.android.state.plugin.internal.StatePlugin
 import io.getstream.chat.android.state.plugin.state.global.GlobalState
 import io.getstream.chat.android.test.TestCoroutineExtension
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -39,9 +37,7 @@ import org.amshove.kluent.`should be`
 import org.amshove.kluent.`should be equal to`
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.RegisterExtension
-import org.mockito.kotlin.any
 import org.mockito.kotlin.doReturn
-import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 import java.util.Date
@@ -127,18 +123,6 @@ internal class MessageComposerControllerTests {
         private val channelState: ChannelState = mock()
         private val globalState: GlobalState = mock()
 
-        init {
-            // Default mocks
-            // Set clientState to initialized
-            whenever(chatClient.awaitInitializationState(any())) doReturn InitializationState.COMPLETE
-            whenever(clientState.initializationState) doReturn MutableStateFlow(InitializationState.COMPLETE)
-            whenever(chatClient.clientState) doReturn clientState
-            // Initialize statePlugin and globalState
-            val statePlugin: StatePlugin = mock()
-            whenever(statePlugin.resolveDependency(eq(GlobalState::class))) doReturn globalState
-            whenever(chatClient.plugins) doReturn listOf(statePlugin)
-        }
-
         fun givenAppSettings(appSettings: AppSettings) = apply {
             whenever(chatClient.getAppSettings()) doReturn appSettings
         }
@@ -149,6 +133,7 @@ internal class MessageComposerControllerTests {
 
         fun givenClientState(user: User) = apply {
             whenever(clientState.user) doReturn MutableStateFlow(user)
+            whenever(chatClient.clientState) doReturn clientState
         }
 
         fun givenChannelState(
@@ -182,6 +167,7 @@ internal class MessageComposerControllerTests {
                 mediaRecorder = mock(),
                 userLookupHandler = mock(),
                 fileToUri = mock(),
+                globalState = MutableStateFlow(globalState),
             )
         }
     }